+++ packages/nimble-components/src/text-field/tests/text-field.stories.ts
@@ -9,6 +9,7 @@ interface TextFieldArgs {
     label: string;
     type: string;
     appearance: string;
+    'clear-inline-padding': boolean;
     value: string;
     readonly: boolean;
     disabled: boolean;
@@ -51,7 +52,7 @@ const metadata: Meta<TextFieldArgs> = {
             type="${x => x.type}"
             appearance="${x => x.appearance}"
             value="${x => x.value}"
-            class="${x => (x.invalid ? 'invalid' : '')}"
+            class="${x => (x.invalid ? 'invalid' : '')} ${x => (x['clear-inline-padding'] ? 'clear-inline-padding' : '')}"
             aria-invalid="${x => x.invalid}"
             ?readonly="${x => x.readonly}"
             ?disabled="${x => x.disabled}"
@@ -78,6 +79,10 @@ const metadata: Meta<TextFieldArgs> = {
             options: Object.values(TextFieldAppearance),
             control: { type: 'radio' }
         },
+        'clear-inline-padding': {
+            description:
+                'Add the class `clear-inline-padding` to remove the start and end padding. Only affects the frameless appearance.'
+        },
         'error-text': {
             description:
                 'A message to be displayed when the text field is in the invalid state explaining why the value is invalid'
@@ -93,6 +98,7 @@ const metadata: Meta<TextFieldArgs> = {
         label: 'default label',
         type: 'text',
         appearance: 'underline',
+        'clear-inline-padding': false,
         value: '',
         readonly: false,
         disabled: false,